Updates to the Previous Version
None of the code or questions have been changed intuitively. The wording has been changed in the questions to make them easier to understand for the students. However, some changes were made in the format to make the document easier for the students to read and to better align the document with the formats used in the notebooks that Ryan have created.